requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
Django==1.8.2
django-ckeditor==5.4.0
django-haystack==2.7.0
django-js-asset==1.0.0
django-materialize-css==0.0.1
django-simditor==0.0.8
Fabric==1.14.0
idna==2.6
jieba==0.39
paramiko==2.4.0
Pillow==5.0.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
requests==2.18.4
six==1.11.0
urllib3==1.22
Whoosh==2.7.4
